Add setup wizard
$ python3 deployment/windows/buildwin32.py
...
$

Fire ISS up, build installer. Result is in Output directory.
Should be around ~4 MB
