+++ deployment/windows/borg-shell.bat
@@ -1,0 +1,4 @@
+SET PATH=%~dp0bin;%PATH%
+SET PYTHONPATH=%~dp0bin
+CD /D %USERPROFILE%
+CMD /K
